[popup] Add option to open extension pages in new tab using keyboard shortcuts (#158)
**Description**
In the current implementation, navigating to the Export and Import pages
through keyboard shortcuts (`Ctrl+Option+I` on Mac or `Ctrl+Alt+I` on
Windows, followed by `E` or `I`) only allows users to open these pages
in the current tab.
Opening these pages in the current tab disrupts the user flow
In order to open the page in a new tab, users have to resort to using a
combination of keyboard commands and mouse clicks which is not ideal.
(control or command + mouse click).

**Enhancement**
This pull request introduces an enhancement to the existing keyboard
shortcut functionalities, allowing users to open the extension pages in
a new tab seamlessly. By holding the Command key (on Mac) or Control key
(on Windows) in conjunction with the existing shortcut keys, users can
now open the extension pages in a new tab, thereby preserving their
position on the current page and facilitating a smoother navigation
experience.

**Testing**
This feature has been tested across different environments to ensure
compatibility and smooth integration with existing functionalities. I
encourage further testing to validate its efficiency and usability
enhancements.

**Request for Review**
I kindly request a review of this pull request and am open to making any
necessary adjustments based on your feedback. Thank you for considering
this valuable enhancement to the user experience.

---------

Co-authored-by: user <user@domain.com>
Co-authored-by: Thomas Prouvot <35368290+tprouvot@users.noreply.github.com>