Feature/improve OAuth flow (#242)
## Describe your changes

**Add default connected app setting and improve access token renew
flow:**
1. There is a now a **default connected app** that can be used for the
situation where the extension cannot use the session Id (when the org
uses API Access Control or Microsoft Defender for Cloud Apps, for
example). This is a big time saver, as it's not needed anymore to create
a connected app from scratch!
2. When the API returns a 401/Unauthorized error, the extension will now
**automatically display a button to renew the access token**. Another
time saver!
3. **Documentation is updated** with the new instructions for setting up
a custom connected app.

Important note: To be able to use a default connected app, the host is
now removed from the redirect Url in the OAuth flow. **This is a ⚠
**BREAKING CHANGE** ⚠ for users who have created their own connected
app**. Those users MUST update their connected app Callback URL to the
new value before they can use this version of the extension. It's
annoying and I'm sorry for it, but I believe it's worth it 😃


## Checklist before requesting a review
- [X] I have read and understand the [Contributions
section](https://github.com/tprouvot/Salesforce-Inspector-reloaded#contributions)
- [X] Target branch is releaseCandidate and not master
- [X] I have performed a self-review of my code
- [X] I ran the [unit
tests](https://github.com/tprouvot/Salesforce-Inspector-reloaded#unit-tests)
and my PR does not break any tests
- [X] I documented the changes I've made on the
[CHANGES.md](https://github.com/tprouvot/Salesforce-Inspector-reloaded/blob/master/CHANGES.md)
and followed actual conventions
- [X] I added a new section on
[how-to.md](https://github.com/tprouvot/Salesforce-Inspector-reloaded/blob/master/docs/how-to.md)
(optional)

---------

Co-authored-by: Mehdi Cherfaoui <>
Co-authored-by: Thomas Prouvot <prouvot.t@gmail.com>